+++ packages/video_player/video_player_android/lib/src/android_video_player.dart
@@ -164,6 +164,11 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
         .setMixWithOthers(MixWithOthersMessage(mixWithOthers: mixWithOthers));
   }
 
+  @override
+  Future<void> setupAudioSession(AudioSession config) {
+    return _api.setupAudioSession(AudioSessionMessage());
+  }
+
   EventChannel _eventChannelFor(int textureId) {
     return EventChannel('flutter.io/videoPlayer/videoEvents$textureId');
   }