+++ packages/video_player/video_player_android/lib/src/messages.g.dart
@@ -191,6 +191,20 @@ class MixWithOthersMessage {
   }
 }
 
+class AudioSessionMessage {
+  AudioSessionMessage();
+
+  Object encode() {
+    final Map<Object?, Object?> pigeonMap = <Object?, Object?>{};
+    return pigeonMap;
+  }
+
+  static AudioSessionMessage decode(Object message) {
+    final Map<Object?, Object?> pigeonMap = message as Map<Object?, Object?>;
+    return AudioSessionMessage();
+  }
+}
+
 class _AndroidVideoPlayerApiCodec extends StandardMessageCodec {
   const _AndroidVideoPlayerApiCodec();
   @override
@@ -215,6 +229,9 @@ class _AndroidVideoPlayerApiCodec extends StandardMessageCodec {
       writeValue(buffer, value.encode());
     } else if (value is VolumeMessage) {
       buffer.putUint8(134);
+      writeValue(buffer, value.encode());
+    } else if (value is AudioSessionMessage) {
+      buffer.putUint8(135);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -244,6 +261,9 @@ class _AndroidVideoPlayerApiCodec extends StandardMessageCodec {
 
       case 134:
         return VolumeMessage.decode(readValue(buffer)!);
+
+      case 135:
+        return AudioSessionMessage.decode(readValue(buffer)!);
 
       default:
         return super.readValueOfType(type, buffer);
@@ -515,6 +535,30 @@ class AndroidVideoPlayerApi {
   Future<void> setMixWithOthers(MixWithOthersMessage arg_msg) async {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.AndroidVideoPlayerApi.setMixWithOthers', codec,
+        binaryMessenger: _binaryMessenger);
+    final Map<Object?, Object?>? replyMap =
+        await channel.send(<Object?>[arg_msg]) as Map<Object?, Object?>?;
+    if (replyMap == null) {
+      throw PlatformException(
+        code: 'channel-error',
+        message: 'Unable to establish connection on channel.',
+      );
+    } else if (replyMap['error'] != null) {
+      final Map<Object?, Object?> error =
+          (replyMap['error'] as Map<Object?, Object?>?)!;
+      throw PlatformException(
+        code: (error['code'] as String?)!,
+        message: error['message'] as String?,
+        details: error['details'],
+      );
+    } else {
+      return;
+    }
+  }
+
+  Future<void> setupAudioSession(AudioSessionMessage arg_msg) async {
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.AndroidVideoPlayerApi.setupAudioSession', codec,
         binaryMessenger: _binaryMessenger);
     final Map<Object?, Object?>? replyMap =
         await channel.send(<Object?>[arg_msg]) as Map<Object?, Object?>?;