+++ packages/video_player/video_player_android/pigeons/messages.dart
@@ -56,6 +56,10 @@ class MixWithOthersMessage {
   bool mixWithOthers;
 }
 
+class AudioSessionMessage {
+  AudioSessionMessage();
+}
+
 @HostApi(dartHostTestHandler: 'TestHostVideoPlayerApi')
 abstract class AndroidVideoPlayerApi {
   void initialize();
@@ -69,4 +73,5 @@ abstract class AndroidVideoPlayerApi {
   void seekTo(PositionMessage msg);
   void pause(TextureMessage msg);
   void setMixWithOthers(MixWithOthersMessage msg);
+  void setupAudioSession(AudioSessionMessage msg);
 }