+++ packages/video_player/video_player_avfoundation/lib/src/avfoundation_video_player.dart
@@ -163,6 +163,11 @@ class AVFoundationVideoPlayer extends VideoPlayerPlatform {
         .setMixWithOthers(MixWithOthersMessage(mixWithOthers: mixWithOthers));
   }
 
+  @override
+  Future<void> setupAudioSession(AudioSession config) {
+    return _api.setupAudioSession(AudioSessionMessage());
+  }
+
   EventChannel _eventChannelFor(int textureId) {
     return EventChannel('flutter.io/videoPlayer/videoEvents$textureId');
   }