+++ packages/video_player/video_player_platform_interface/lib/method_channel_video_player.dart
@@ -149,6 +149,12 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
     );
   }
 
+  @override
+  Future<void> setupAudioSession(AudioSession config) {
+    final msg = AudioSessionMessage();
+    return _api.setupAudioSession(msg);
+  }
+
   EventChannel _eventChannelFor(int textureId) {
     return EventChannel('flutter.io/videoPlayer/videoEvents$textureId');
   }