+++ packages/video_player/video_player_platform_interface/pigeons/messages.dart
@@ -42,6 +42,8 @@ class MixWithOthersMessage {
   bool mixWithOthers;
 }
 
+class AudioSessionMessage {}
+
 @HostApi(dartHostTestHandler: 'TestHostVideoPlayerApi')
 abstract class VideoPlayerApi {
   void initialize();
@@ -55,6 +57,7 @@ abstract class VideoPlayerApi {
   void seekTo(PositionMessage msg);
   void pause(TextureMessage msg);
   void setMixWithOthers(MixWithOthersMessage msg);
+  void setupAudioSession(AudioSessionMessage msg);
 }
 
 void configurePigeon(PigeonOptions opts) {