+++ packages/video_player/video_player_platform_interface/test/video_player_options_test.dart
@@ -20,4 +20,11 @@ void main() {
       expect(options.mixWithOthers, false);
     },
   );
+  test(
+    'VideoPlayerOptions bypassAudioSetup defaults to false',
+    () {
+      final VideoPlayerOptions options = VideoPlayerOptions();
+      expect(options.bypassAudioSetup, false);
+    },
+  );
 }