+++ packages/video_player/video_player_web/lib/video_player_web.dart
@@ -146,4 +146,8 @@ class VideoPlayerPlugin extends VideoPlayerPlatform {
   /// Sets the audio mode to mix with other sources (ignored)
   @override
   Future<void> setMixWithOthers(bool mixWithOthers) => Future<void>.value();
+
+  ///
+  @override
+  Future<void> setupAudioSession(AudioSession conf) => Future<void>.value();
 }